Remove NVDA browse mode adjustment
Co-authored-by: Chris Yoong chris-gds@users.noreply.github.com

Remove the CSS adjustment that made the NVDA browse mode perceive the button content as one element. Whilst this was potentially an improvement to how NVDA users perceived the button, we found in testing that it caused long announcements to be cut off. More details of the initial fix here: https://github.com/alphagov/govuk-frontend/pull/2257/commits/6eb5bf788d3fff5f0e11626637df9cad95839fab

The reason for the announcement being cut off appears to be related to the character limit for line length that NVDA users can set for the browse mode. The issue was that whilst the browse mode default behaviour is to read out the subsequent line(s) when user presses down again (so that eventually the user will hear all of the content), here those subsequent lines were not being read out - the reason for this could be related to the amount of content inside the button and NVDA perceiving that there was more content (namely the button element itself) to announce.

@chris-gds has done a more thorough write up of the issue of the NVDA announcements being cut off as it was observed in: https://github.com/alphagov/govuk-frontend/issues/2373#issuecomment-931544190